website/requirements.txt
flask
pinecone-client
sentence-transformers
python-dotenv
flask-limiter
redis
validators
beautifulsoup4
requests
diskcache
gunicorn
validators
qdrant-client
